fix more compiles

+++ co_async/generic/io_context.hpp
@@ -45,7 +45,7 @@ public:
                    std::span<IOContext> peerContexts) {
         IOContextGuard guard(this);
         if (options.threadAffinity) {
-            PlatformIOContext::schedSetThreadAffinity(*options.threadAffinity);
+            PlatformIOContext::schedSetThreadAffinity((int)*options.threadAffinity);
         }
         auto maxSleep = options.maxSleep;
         while (!stop.stop_requested()) [[likely]] {
